fix(telemetry): address review P1/P2 issues in PR #915
- Guard CREATED_ISSUE= backfill behind execution phase + auto-fix source
  check (phase_str == "execution") to prevent reviewer/planner output
  from corrupting external_id
- Extract streaming helpers to helpers/streaming.rs submodule so
  helpers.rs stays under the 800-line limit
- Set ExecutionPhase::SimpleReview on triage check_req in review_loop.rs
  so review rounds never trigger the execution-phase backfill gate
- Defer triage "completed" observability log until after parse_triage
  succeeds; emit "failed" with TurnFailureKind::Protocol on parse error
  so the DB never shows a completed triage turn that produced no decision
- Extract triage_pipeline tests to triage_pipeline_tests.rs to keep
  triage_pipeline.rs under 800 lines after the deferred-log restructure
